+++ src/inc/hc/hc_edf.h
@@ -89,7 +89,7 @@ typedef struct {
     ocrTask_t base;
     reg_node_t * waiters;
     reg_node_t * signalers; // Does not grow, set once when the task is created
-    size_t nbdeps;
+    u64 nbdeps;
     ocrEdt_t p_function;
 } ocrTaskHc_t;
